+++ tests/package-common/src/test-adapter.ts
@@ -1,5 +1,6 @@
 import {
   type Adapter,
+  type AdapterArgs,
   ClientSessionLeaderThreadProxy,
   type LockStatus,
   type MakeSqliteDb,
@@ -69,7 +70,7 @@ export const makeTestAdapter = ({
 } = {}): Adapter =>
   ((adapterArgs) =>
     Effect.gen(function* () {
-      const { schema, storeId, syncPayloadEncoded, syncPayloadSchema } = adapterArgs
+      const { schema, storeId, syncPayloadEncoded, syncPayloadSchema, params } = adapterArgs
       const sqlite3 = yield* Effect.promise(() => loadSqlite3Wasm())
       const makeSqliteDb = yield* sqliteDbFactory({ sqlite3 })
       const shutdownChannel = yield* makeShutdownChannel(storeId)
@@ -94,6 +95,7 @@ export const makeTestAdapter = ({
         syncOptions: sync,
         syncPayloadEncoded,
         syncPayloadSchema,
+        params,
         testing,
         shutdownChannel,
       }).pipe(UnknownError.mapToUnknownError)
@@ -130,6 +132,7 @@ const makeLocalLeaderThread = ({
   syncOptions,
   syncPayloadEncoded,
   syncPayloadSchema,
+  params,
   testing,
   shutdownChannel,
 }: {
@@ -140,6 +143,7 @@ const makeLocalLeaderThread = ({
   syncOptions: SyncOptions | undefined
   syncPayloadEncoded: Schema.Json | undefined
   syncPayloadSchema: Schema.Decoder<Schema.Json> | undefined
+  params: AdapterArgs['params']
   testing?: { overrides?: TestingOverrides }
   shutdownChannel: ShutdownChannel.ShutdownChannel
 }) =>
@@ -175,6 +179,7 @@ const makeLocalLeaderThread = ({
         shutdownChannel,
         syncPayloadEncoded,
         syncPayloadSchema,
+        params,
       }).pipe(Layer.provide(StateHead.layer({ dbState }))),
     )